Various updates to advanced config
I have widened the window so it no longer needs to be stretched to see Map info. I have fixed new ID validation so you can't make an object with no ID. Additionally I have begun the process of renaming functions and adding comments to be compliant with our coding and commenting conventions we have outlined.